+++ avocado/plugins/nrun.py
@@ -4,7 +4,8 @@ import os
 import random
 import sys
 
-from avocado.core import exit_codes, nrunner, parser_common_args, resolver
+from avocado.core import (exit_codes, nrunner, parser_common_args, resolver,
+                          status_server)
 from avocado.core.future.settings import settings
 from avocado.core.output import LOG_UI
 from avocado.core.parser import HintParser
@@ -151,10 +152,11 @@ class NRun(CLICmd):
 
             listen = config.get('nrun.status_server.listen')
             verbose = config.get('core.verbose')
-            self.status_server = nrunner.StatusServer(listen,  # pylint: disable=W0201
-                                                      [t.identifier for t in
-                                                       self.pending_tasks],
-                                                      verbose)
+            self.status_server = status_server.StatusServer(
+                listen,  # pylint: disable=W0201
+                [t.identifier for t in
+                 self.pending_tasks],
+                verbose)
             self.status_server.start()
             parallel_tasks = config.get('nrun.parallel_tasks')
             loop = asyncio.get_event_loop()
